[1] fix the blocking error : Add <queries>: Add the <queries> tag to ProviderTest's AndroidManifest.xml . Add <package android:name="com.example.databasetest" />: This is the crucial part. It tells the system that ProviderTest needs to be able to interact with apps in the com.example.databasetest package.
Thanks to Stackoverflow :
Qte ----- Apps targeting API level 30 and above (Android 11+) need to specify which other packages they need to see (or use) on the device. For this, developers need to add a <queries> tag in their apps' AndroidManifest.xml --- Unqte
and also Qte ----- The <package> tag is one of the valid tags under <queries> ... Without the <queries> tag, AppsFilter will not allow access of any package to your app, which is what you're seeing in your logcat. --- Unqte

Note : The installation order is important: We have to have DatabaseTest (the provider app) installed (built, up and running) first.